feat: finish modules task
